Restructure the code such that it can be compiled with nvcc 11.2
